refactor(#5031): extract delivery lease cell impl
